Move wrapper to the dataservice-write
Move MultipartUploadWrapper to the
`dataservice-write`.

Remove unneeded anymore the `TypedEvent` class.
It was used in the `MultipartUploadWrapper`
in the first stages.

Resolves: OLPEDGE-2456
