+++ @here/olp-sdk-dataservice-write/lib/utils/BlobData.ts
