+++ @here/olp-sdk-dataservice-write/lib/utils/multipartupload-internal/BlobV1UploadRequest.ts
