+++ @here/olp-sdk-dataservice-write/lib/utils/multipartupload-internal/NodeFileData.ts
@@ -17,7 +17,8 @@
  * License-Filename: LICENSE
  */
 
-import { BlobData, fs } from "@here/olp-sdk-core";
+import { fs } from "@here/olp-sdk-core";
+import { BlobData } from "@here/olp-sdk-dataservice-write";
 
 /**
  * @internal
