+++ @here/olp-sdk-dataservice-write/lib/utils/multipartupload-internal/UploadRequest.ts
