+++ @here/olp-sdk-dataservice-write/test/unit/BufferData.test.ts
