+++ @here/olp-sdk-dataservice-write/test/unit/NodeFileData.test.ts
